add permission restrictions and pagination

+++ app/Http/Controllers/HomeController.php
@@ -2,10 +2,11 @@
 
 namespace App\Http\Controllers;
 
-
+use App\Enums\PermissionEnum;
 use App\Enums\RoleEnum;
 use App\Models\Role;
 use App\Repositories\UserRepository;
+use App\Services\PermissionService;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Auth;
 
@@ -16,7 +17,7 @@ class HomeController extends Controller
      *
      * @return void
      */
-    public function __construct(private UserRepository $userRepository)
+    public function __construct(private UserRepository $userRepository, private PermissionService $permissionService)
     {
         $this->middleware('auth');
     }
@@ -43,15 +44,27 @@ class HomeController extends Controller
 
     public function show(Request $request)
     {
+        $userPermissions = $this->permissionService->getPermissions();
+        
         $client = $this->userRepository->findClientByUuid($request->uuid)->first();
+        
+        if (Auth::user()?->client_uuid === $client->client_uuid || in_array(PermissionEnum::Edit->value, $userPermissions)) {
+            return view('client', ['client' => $client, 'userPermissions' => $userPermissions]);
+        }
+
+        return redirect(route('home'));
    
-        return view('client', ['client' => $client]);
     }
 
     public function update(Request $request)
     {
-        $client = $this->userRepository->findClientByUuid($request->uuid)->first();
+        $hasPermissions = $this->permissionService->checkPermissions(PermissionEnum::Edit->value);
 
+        if (! $hasPermissions) {
+            return redirect(route('home'));
+        }
+
+        $client = $this->userRepository->findClientByUuid($request->uuid)->first();
         $contactPeople = $client->contactPerson->toArray();
 
         return view('editclient', ['client' => $client, 'contactPeople' => $contactPeople]);
@@ -60,22 +73,29 @@ class HomeController extends Controller
 
     public function edit(Request $request)
     {
-        $data = [];
-        
-        $data['client_uuid'] = $request->uuid;
-        $data['name'] = $request->name;
-        $data['address'] = $request->address;
-        $data['email'] = $request->email;
-       
-        $this->userRepository->update($data);
+        $hasPermissions = $this->permissionService->checkPermissions(PermissionEnum::Edit->value);
+
+        if ($hasPermissions) {
+            $data = [];
+            $data['client_uuid'] = $request->uuid;
+            $data['name'] = $request->name;
+            $data['address'] = $request->address;
+            $data['email'] = $request->email;
+           
+            $this->userRepository->update($data);
+        }
 
         return redirect(route('home'));
     }
 
     public function delete(int $id)
     {
-        $this->userRepository->delete($id);
+        $hasPermissions = $this->permissionService->checkPermissions(PermissionEnum::Delete->value);
 
+        if ($hasPermissions) {
+            $this->userRepository->delete($id);
+        }
+      
         return redirect(route('home'));
     }
 }
